requirements.txt
pysros
python-dotenv